refactor: decompose update-cli.ts into modular components
Extract 6 modules from 885-line monolith:
- channel-resolver.ts — npm dist-tag/version resolution
- registry-client.ts — credential redaction
- version-comparator.ts — semver comparison + dev-channel detection
- package-manager-runner.ts — exec update + verify install
- post-update-handler.ts — kit init + migrate pipeline
- types.ts + error.ts — shared types and error class

update-cli.ts is now a thin orchestrator (~100 lines logic).
All 100 tests pass, zero behavioral changes.

Closes #606
Part of #600
